+++ src/cmd/servicePush.go
@@ -126,8 +126,6 @@ func servicePushCmd() *cmdBuilder.Cmd {
 				return err
 			}
 
-			uxBlocks.PrintInfo(styles.InfoLine(i18n.T(i18n.PushDeployCreatingPackageDone)))
-
 			if cmdData.Params.GetString("archiveFilePath") != "" {
 				uxBlocks.PrintInfo(styles.InfoLine(i18n.T(i18n.PushDeployPackageSavedInto, cmdData.Params.GetString("archiveFilePath"))))
 			}